further refactored file structure and namespaces
